+++ .build/tasks/Build-Module.ModuleBuilder.build.ps1
@@ -4,7 +4,7 @@ Param (
     [string]
     $ProjectName = (property ProjectName $(
             #Find the module manifest to deduce the Project Name
-            (Get-ChildItem $BuildRoot\*\*.psd1 | Where-Object {
+            (Get-ChildItem $BuildRoot\*\*.psd1 -Exclude 'build.psd1', 'analyzersettings.psd1' | Where-Object {
                     ($_.Directory.Name -match 'source|src' -or $_.Directory.Name -eq $_.BaseName) -and
                     $(try
                         {
@@ -12,6 +12,7 @@ Param (
                         }
                         catch
                         {
+                            Write-Warning $_
                             $false
                         }) }
             ).BaseName
@@ -20,7 +21,7 @@ Param (
 
     [Parameter()]
     [string]
-    $SourcePath = (property SourcePath ((Get-ChildItem $BuildRoot\*\*.psd1 | Where-Object {
+    $SourcePath = (property SourcePath ((Get-ChildItem $BuildRoot\*\*.psd1 -Exclude 'build.psd1', 'analyzersettings.psd1' | Where-Object {
                     ($_.Directory.Name -match 'source|src' -or $_.Directory.Name -eq $_.BaseName) -and
                     $(try
                         {
@@ -28,6 +29,7 @@ Param (
                         }
                         catch
                         {
+                            Write-Warning $_
                             $false
                         }) }
             ).Directory.FullName)
@@ -258,7 +260,8 @@ Task Build_NestedModules_ModuleBuilder {
                             catch
                             {
                                 $false
-                            }) }
+                            })
+                    }
             ).FullName -replace [Regex]::Escape($BuiltModuleBase), ".$([io.path]::DirectorySeparatorChar)"