+++ CHANGELOG.md
@@ -7,7 +7,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Added warning messages to all build task if the task couldn't be imported because of an invalid PSD1 file
 -  `build.ps1` will now dynamically determine the build configuration if not specified via the `-BuildConfig` parameter
+
+## [0.100.0] - 2020-02-01
+
+### Added
+
+- Added the option to specify `CodeCoverageOutputFile` and `CodeCoverageOutputFileEncoding`
+  in the file `build.yml`. For example if a code coverage provider need the
+  file to be named in a certain way.
+
+### Changed
+
+- Added new common functions for build tasks to reduce code duplication.
 
 ## [0.99.4] - 2020-01-22